refactor user_guid_test
